Update README.md
My team is participating in the "SpaceTrash Hack: Revolutionizing Recycling on Mars" challenge, and we have some questions regarding our project proposal.

We are considering a system that uses supercritical carbon dioxide (SC-CO2) for cleaning fabrics and materials. Additionally, we would like to know if it is allowed to combine plastic waste with Mars regolith to create base materials for containers or other reusable products. 
Finally, we would like clarification on whether we need to propose the desing of tools and systems for managing and processing recycled materials as part of the project.

Could you please confirm it these aprroaches are acceptables within the scope of the challenge? We want to ensure thet our proposal aligns with the rules and objetives of the SpaceTrash Hack.